God class is using a class Game to play. Use DebugMatter to show extra data for debugging.

+++ src/com/company/NodeInterface.java
@@ -14,5 +14,5 @@ public interface NodeInterface {
 
     double getCenterOfMassPosY();
 
-    float getMass();
+    long getMass();
 }
